requirements.txt
cachetools==2.1.0
certifi==2018.10.15
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.2
django-oauth-toolkit==1.2.0
django-rest-registration==0.3.13
djangorestframework==3.9.0
google-api-core==1.5.1
google-auth==1.5.1
google-cloud-core==0.28.1
google-cloud-speech==0.36.0
google-cloud-storage==1.13.0
google-resumable-media==0.3.1
googleapis-common-protos==1.5.5
grpcio==1.16.0
idna==2.7
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
oauthlib==2.1.0
protobuf==3.6.1
psycopg2==2.7.5
pyasn1==0.4.4
pyasn1-modules==0.2.2
python-decouple==3.1
pytz==2018.6
requests==2.20.0
rsa==4.0
six==1.11.0
uritemplate==3.0.0
urllib3==1.24
